+++ cacus/__init__.py
@@ -1,4 +1,3 @@
 #!/usr/bin/env python
 
 from .common import *
-import cli